// src/env.enum.ts
export enum EnvEnum {
  NODE_ENV = 'NODE_ENV',

  I18N_MULTILINGUAL_ENABLED = "I18N_MULTILINGUAL_ENABLED",
  I18N_DEFAULT_LANG = "I18N_DEFAULT_LANG",

  SITE_BRAND_NAME = "SITE_BRAND_NAME",
  SITE_PROTOCOL = "SITE_PROTOCOL",
  SITE_HOST = "SITE_HOST",
  SITE_PORT = "SITE_PORT",
  SITE_BASE_URL = "SITE_BASE_URL",
  SITE_CORS_ORIGINS = "SITE_CORS_ORIGINS",

  COOKIE_SECRET = "COOKIE_SECRET",

  AUTH_ACCESS_TOKEN_SECRET = "AUTH_ACCESS_TOKEN_SECRET",
  AUTH_ACCESS_TOKEN_EXPIRATION = "AUTH_ACCESS_TOKEN_EXPIRATION",
  AUTH_REFRESH_TOKEN_SECRET = "AUTH_REFRESH_TOKEN_SECRET",
  AUTH_REFRESH_TOKEN_EXPIRATION = "AUTH_REFRESH_TOKEN_EXPIRATION",

  OAUTH_GOOGLE_CLIENT_ID = "OAUTH_GOOGLE_CLIENT_ID",
  OAUTH_GOOGLE_SECRET = "OAUTH_GOOGLE_SECRET",
  OAUTH_GOOGLE_CALLBACK_URL = "OAUTH_GOOGLE_CALLBACK_URL",

  DB_TYPE = "DB_TYPE",
  DB_HOST = "DB_HOST",
  DB_PORT = "DB_PORT",
  DB_USERNAME = "DB_USERNAME",
  DB_PASSWORD = "DB_PASSWORD",
  DB_NAME = "DB_NAME",

  REDIS_URL = "REDIS_URL",
  REDIS_HOST = "REDIS_HOST",
  REDIS_PORT = "REDIS_PORT",
  REDIS_PASSWORD = "REDIS_PASSWORD",

  MAILER_HOST = "MAILER_HOST",
  MAILER_PORT = "MAILER_PORT",
  MAILER_SECURE = "MAILER_SECURE",
  MAILER_USER = "MAILER_USER",
  MAILER_PASSWORD = "MAILER_PASS",

  SMS_PROVIDER = "SMS_PROVIDER",
  SMS_ORIGINATOR = "SMS_ORIGINATOR",
  SMS_API_KEY = "SMS_API_KEY",

  UPLOAD_ROUTE = "UPLOAD_ROUTE",

  S3_ENDPOINT = "S3_ENDPOINT",
  S3_REGION = "S3_REGION",
  S3_ACCESS_KEY = "S3_ACCESS_KEY",
  S3_SECRET_KEY = "S3_SECRET_KEY",
  S3_BUCKET = "S3_BUCKET",
  S3_ALLOWED_ORIGINS = "S3_ALLOWED_ORIGINS",
  S3_UPLOAD_LINK_EXPIRATION_IN_SECONDS = "S3_UPLOAD_LINK_EXPIRATION_IN_SECONDS",
  S3_ACL = "S3_ACL",
  S3_FORTH_PATH_STYLE = "S3_FORTH_PATH_STYLE",
  S3_STREAMING_UPLOAD = "S3_STREAMING_UPLOAD",

  COMMENT_REPLY_ALLOWED_LEVEL = "COMMENT_REPLY_ALLOWED_LEVEL"
}